cricao da tela soundModalView
